Restructure Android build with a Dockerfile.
Remove the "works on my machine" guesswork. This keeps all the Android
parts of the build in build.sh, but moves all the host dependencies
(including the capnproto fetch) into the Dockerfile.

This cleans up some of the header/library directory handling as well,
which was necessary for making the build script work when the host and
target builds are separate (CAPNP_CFLAGS doesn't actually exist, so
the build was previously finding the host's includes by coincidence?)