Fix rr build for newer NDKs.
The presence of the execinfo.h header does not mean that the backtrace
APIs will be available at runtime on Android. The header is always
available (to facilitate runtime detection of APIs via `dlsym`), but
the API is only available for devices that are API 33 or newer.

This was not symptomatic with older NDKs because the execinfo.h header
was not yet present (it was added in r24 or r25; rr was tested for
Android with r23).